Add disable_print and disable_overlay envvars

+++ print.sh
@@ -1,3 +1,8 @@
+if [ "${DISABLE_PRINT}" = "true" ]; then
+    echo "Print disabled, exiting early."
+    exit 0
+fi
+
 rm -rf print.pdf
 convert tiled_photos.jpg print.pdf
 lp -d EPSON_ET_2850_Series print.pdf
